refactor: 更新路由表配置
# Conflicts:
#	src/BootstrapBlazor.Shared/docs.json
