+++ src/BootstrapBlazor.Shared/Demos/Speeches/Recognizers/RecognizerNormal.razor
