+++ src/BootstrapBlazor.Shared/Demos/Speeches/SpeechWaves/SpeechWaveNormal.razor
