+++ src/BootstrapBlazor.Shared/Demos/Speeches/SpeechWaves/SpeechWaveShow.razor
@@ -1,4 +1,4 @@
-﻿@inject IStringLocalizer<SpeechWaves> Localizer
+﻿@inject IStringLocalizer<SpeechWaveShow> Localizer
 
 <div class="row g-3">
     <div class="col-12">
