+++ src/BootstrapBlazor.Shared/Demos/Speeches/SpeechWaves/SpeechWaveShowUsedTime.razor
