+++ src/BootstrapBlazor.Shared/Demos/Speeches/Synthesizers/SynthesizerNormal.razor
