+++ src/BootstrapBlazor.Shared/Samples/Bluetooths.razor
@@ -1,5 +1,4 @@
 ﻿@page "/blue-tooth"
-@namespace BootstrapBlazor.Shared.Samples
 @inject IStringLocalizer<Bluetooths> Localizer
 
 <h3>@Localizer["Title"]</h3>
