+++ src/BootstrapBlazor.Shared/Samples/Bluetooths.razor.cs
