+++ src/BootstrapBlazor.Shared/Samples/Bluetooths.razor.css
