+++ src/BootstrapBlazor.Shared/Samples/DockViews.razor
@@ -1,4 +1,4 @@
-﻿@page "/dockview"
+﻿@page "/dock-view"
 @inject IStringLocalizer<DockViews> Localizer
 
 <h3>@Localizer["Title"]</h3>
