+++ src/BootstrapBlazor.Shared/Samples/SignaturePadPageResponsive.razor
@@ -1,6 +1,5 @@
 ﻿@page "/signature-responsive"
 @inject IStringLocalizer<SignaturePads> Localizer
-@namespace BootstrapBlazor.Shared.Samples
 
 <h3>@Localizer["ResponsiveInterfaceText"]</h3>
