+++ src/BootstrapBlazor.Shared/Samples/SignaturePads.razor
@@ -1,5 +1,4 @@
 ﻿@page "/signature-pad"
-@namespace BootstrapBlazor.Shared.Samples
 @inject IStringLocalizer<SignaturePads> Localizer
 
 <h3>@Localizer["Title"]</h3>
