+++ src/BootstrapBlazor.Shared/Samples/SignaturePads.razor.cs
