+++ src/BootstrapBlazor.Shared/Samples/Speeches/Index.razor
@@ -1,5 +1,4 @@
-﻿@page "/speech"
-@namespace BootstrapBlazor.Shared.Samples
+﻿@page "/speech/index"
 @inherits BootstrapComponentBase
 @layout MainLayout
