+++ src/BootstrapBlazor.Shared/Samples/Speeches/Recognizers.razor
@@ -1,5 +1,4 @@
-﻿@page "/recognizer"
-@namespace BootstrapBlazor.Shared.Samples
+﻿@page "/speech/recognizer"
 @inherits BootstrapComponentBase
 @inject IStringLocalizer<Recognizers> Localizer
 
@@ -9,7 +8,7 @@
 <DemoBlock Title="@Localizer["RecognizerNormalTitle"]"
            Introduction="@Localizer["RecognizerNormalIntro"]"
            Name="Normal"
-           Demo="typeof(Demos.Speechs.Recognizers.RecognizerNormal)">
+           Demo="typeof(Demos.Speeches.Recognizers.RecognizerNormal)">
     <p><b>@Localizer["RecognizerNormalDescription"]：</b></p>
     <p>@((MarkupString)Localizer["RecognizerNormalTips"].Value)</p>
 </DemoBlock>
