+++ src/BootstrapBlazor.Shared/Samples/Speeches/SpeechWaves.razor.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Website: https://www.blazor.zone or https://argozhang.github.io/
 
-namespace BootstrapBlazor.Shared.Samples;
+namespace BootstrapBlazor.Shared.Samples.Speeches;
 
 /// <summary>
 /// 语音识别波形图示例
@@ -15,29 +15,32 @@ public partial class SpeechWaves
     /// <returns></returns>
     private IEnumerable<AttributeItem> GetAttributes() => new AttributeItem[]
     {
-
-        new() {
+        new()
+        {
             Name = nameof(SpeechWave.Show),
             Description = Localizer["ShowAttr"],
             Type = "bool",
             ValueList = "true/false",
             DefaultValue = "false"
         },
-        new() {
+        new()
+        {
             Name = nameof(SpeechWave.ShowUsedTime),
             Description = Localizer["ShowUsedTimeAttr"],
             Type = "bool",
             ValueList = "true/false",
             DefaultValue = "true"
         },
-        new() {
+        new()
+        {
             Name = nameof(SpeechWave.OnTimeout),
             Description = Localizer["OnTimeoutAttr"],
             Type = "Func<Task>",
             ValueList = " — ",
             DefaultValue = " — "
         },
-        new() {
+        new()
+        {
             Name = nameof(SpeechWave.TotalTime),
             Description = Localizer["TotalTimeSecondAttr"],
             Type = "int",
