+++ src/BootstrapBlazor.Shared/Samples/Steps.razor
@@ -1,4 +1,4 @@
-﻿@page "/Step"
+﻿@page "/step"
 @inject IStringLocalizer<Steps> Localizer
 
 <h3>@Localizer["StepsTitle"]</h3>
