+++ src/BootstrapBlazor.Shared/docs.json
@@ -1,4 +1,182 @@
 ﻿{
+  "src": {
+    "ajax": "Ajaxs",
+    "alert": "Alerts",
+    "anchor-link": "AnchorLinks",
+    "anchor": "Anchors",
+    "auto-complete": "AutoCompletes",
+    "auto-fill": "AutoFills",
+    "auto-redirect": "AutoRedirects",
+    "avatar": "Avatars",
+    "badge": "Badges",
+    "ocr": "BaiduOcr",
+    "barcode-reader": "BarcodeReaders",
+    "block": "Blocks",
+    "blue-tooth": "Bluetooths",
+    "breadcrumb": "Breadcrumbs",
+    "button": "Buttons",
+    "calendar": "Calendars",
+    "camera": "Cameras",
+    "captcha": "Captchas",
+    "card": "Cards",
+    "carousel": "Carousels",
+    "cascader": "Cascaders",
+    "checkbox-list": "CheckboxLists",
+    "checkbox": "Checkboxs",
+    "cherry-markdown": "CherryMarkdowns",
+    "circle": "Circles",
+    "client": "Client",
+    "collapse": "Collapses",
+    "color-picker": "ColorPickers",
+    "console": "Consoles",
+    "context-menu": "ContextMenus",
+    "count-up": "CountUps",
+    "datetime-picker": "DateTimePickers",
+    "datetime-range": "DateTimeRanges",
+    "dialog": "Dialogs",
+    "dispatch": "Dispatches",
+    "display": "Displays",
+    "divider": "Dividers",
+    "download": "Downloads",
+    "drag-drop": "DragDrops",
+    "drawer": "Drawers",
+    "dropdown": "Dropdowns",
+    "dropdown-widget": "DropdownWidgets",
+    "edit-dialog": "EditDialogs",
+    "editor-form": "EditorForms",
+    "editor": "Editors",
+    "empty": "Empties",
+    "eye-dropper": "EyeDroppers",
+    "fa-icon": "FAIcons",
+    "file-icon": "FileIcons",
+    "file-viewer": "FileViewers",
+    "floating-label": "FloatingLabels",
+    "footer": "Footers",
+    "geolocation": "Geolocations",
+    "global-exception": "GlobalException",
+    "go-top": "GoTops",
+    "group-box": "GroupBoxes",
+    "handwritten": "Handwrittens",
+    "html-renderer": "HtmlRenderers",
+    "image-viewer": "ImageViewers",
+    "input-number": "InputNumbers",
+    "input": "Inputs",
+    "ip": "Ips",
+    "layout-page": "LayoutPages",
+    "layout": "Layouts",
+    "light": "Lights",
+    "link-button": "LinkButtons",
+    "list-view": "ListViews",
+    "live2d-display": "Live2DDisplays",
+    "locator": "Locators",
+    "logout": "Logouts",
+    "markdown": "Markdowns",
+    "menu": "Menus",
+    "message": "Messages",
+    "modal": "Modals",
+    "mouse-follower": "MouseFollowers",
+    "multi-select": "MultiSelects",
+    "navigation": "Navigation",
+    "notification": "Notifications",
+    "onscreen-keyboard": "OnScreenKeyboards",
+    "pagination": "Paginations",
+    "pdf-reader": "PdfReaders",
+    "pop-confirm": "PopoverConfirms",
+    "popover": "Popovers",
+    "print-view": "PrintView",
+    "progress": "Progress",
+    "qr-code": "QRCodes",
+    "radio": "Radios",
+    "rate": "Rates",
+    "reconnector": "Reconnectors",
+    "repeater": "Repeaters",
+    "responsive": "Responsives",
+    "ribbon-tab": "RibbonTabs",
+    "row": "Rows",
+    "scroll": "Scrolls",
+    "search-dialog": "SearchDialogs",
+    "search": "Searches",
+    "select": "Selects",
+    "select-tree": "SelectTrees",
+    "signature-pad": "SignaturePads",
+    "skeleton": "Skeletons",
+    "spinner": "Spinners",
+    "split": "Splits",
+    "step": "Steps",
+    "sweet-alert": "SweetAlerts",
+    "switch-button": "SwitchButtons",
+    "switch": "Switches",
+    "tab": "Tabs",
+    "tag": "Tags",
+    "textarea": "TextAreas",
+    "timeline": "Timelines",
+    "timer": "Timers",
+    "toast": "Toasts",
+    "toggle": "Toggles",
+    "tooltip": "Tooltips",
+    "topology": "Topologies",
+    "transfer": "Transfers",
+    "transition": "Transitions",
+    "tree": "Trees",
+    "tree-view": "TreeViews",
+    "upload": "Uploads",
+    "validate-form": "ValidateForms",
+    "video-player": "VideoPlayers",
+    "dock-view": "DockViews",
+    "fullscreen": "FullScreens",
+    "html2pdf": "Html2Pdfs",
+    "input-group": "InputGroups",
+    "label": "Labels",
+    "layout-demo": "LayoutDemo",
+    "layout-page1": "LayoutPages1",
+    "print": "Print",
+    "pulse-button": "PulseButtons",
+    "signature-responsive": "SignaturePadPageResponsive",
+    "title": "Titles",
+    "slider": "Sliders",
+    "table": "Table\\Tables",
+    "table/cell": "Table\\TablesCell",
+    "table/auto-refresh": "Table\\TablesAutoRefresh",
+    "table/column": "Table\\TablesColumn",
+    "table/column/drag": "Table\\TablesColumnDrag",
+    "table/column/list": "Table\\TablesColumnList",
+    "table/column/resizing": "Table\\TablesColumnResizing",
+    "table/column/template": "Table\\TablesColumnTemplate",
+    "table/detail": "Table\\TablesDetailRow",
+    "table/dialog": "Table\\TablesDialog",
+    "table/dynamic": "Table\\TablesDynamic",
+    "table/dynamic-excel": "Table\\TablesDynamicExcel",
+    "table/dynamic-object": "Table\\TablesDynamicObject",
+    "table/edit": "Table\\TablesEdit",
+    "table/excel": "Table\\TablesExcel",
+    "table/export": "Table\\TablesExport",
+    "table/filter": "Table\\TablesFilter",
+    "table/fix-column": "Table\\TablesFixedColumn",
+    "table/header": "Table\\TablesFixedHeader",
+    "table/footer": "Table\\TablesFooter",
+    "table/multi-header": "Table\\TablesHeader",
+    "table/loading": "Table\\TablesLoading",
+    "table/lookup": "Table\\TablesLookup",
+    "table/page": "Table\\TablesPages",
+    "table/row": "Table\\TablesRow",
+    "table/search": "Table\\TablesSearch",
+    "table/selection": "Table\\TablesSelection",
+    "table/toolbar": "Table\\TablesToolbar",
+    "table/tracking": "Table\\TablesTracking",
+    "table/tree": "Table\\TablesTree",
+    "table/virtualization": "Table\\TablesVirtualization",
+    "table/wrap": "Table\\TablesWrap",
+    "speech/index": "Speeches\\Index",
+    "speech/wave": "Speeches\\SpeechWaves",
+    "speech/recognizer": "Speeches\\Recognizers",
+    "speech/synthesizer": "Speeches\\Synthesizers",
+    "chart/index": "Charts\\Index",
+    "chart/bar": "Charts\\Bar",
+    "chart/bubble": "Charts\\Bubble",
+    "chart/doughnut": "Charts\\Doughnut",
+    "chart/line": "Charts\\Line",
+    "chart/pie": "Charts\\Pie"
+  },
   "video": {
     "autorefresh": "BV1ap4y1x7Qn?p=8",
     "column": "BV1ap4y1x7Qn?p=2",
@@ -42,5 +220,15 @@
     "reconnectors": "BV1Dr4y1J7Z5;BV193411P7Dz",
     "speechwaves": "BV1Dr4y1J7Z5",
     "topologies": "BV1eY4y167jn;BV13Z4y1h7MA"
+  },
+  "link": {
+    "Ant design of blazor": "https://ant-design-blazor.github.io/",
+    "Bulma": "https://gitee.com/loogn/bulmarazor",
+    "CCFlow": "http://ccflow.org/?from=blazor",
+    "Dotnet9": "https://dotnet9.com/",
+    "Lin Dexi's Blog": "https://blog.lindexi.com",
+    "M&J": "http://www.mrhuo.com/",
+    "Pear Admin": "http://www.pearadmin.com",
+    "SAPHP": "https://www.swiftadmin.net/"
   }
 }
